# Conflicts:
#	modular/code/game/objects/tent_kit.dm

+++ code/game/objects/structures/traps.dm
@@ -17,7 +17,7 @@
 	var/used_time = 14 // interaction time for disabling traps, scales down with trap skill
 
 
-	var/list/static/ignore_typecache
+	var/static/list/ignore_typecache
 	var/list/mob/immune_minds = list() //unused and a bit weird, helpful for making mobs immune to the traps without TRAIT_LIGHT_STEP
 
 	var/sparks = TRUE
@@ -160,6 +160,12 @@
 	time_between_triggers = 10
 	flare_message = span_warning("[src] snaps shut!")
 
+/obj/structure/trap/stun/hunter/Destroy()
+	if(stored_item && !QDELING(stored_item))
+		qdel(stored_item)
+	stored_item = null
+	return ..()
+
 /obj/structure/trap/stun/hunter/Crossed(atom/movable/AM)
 	caught = TRUE
 	. = ..()
@@ -203,7 +209,9 @@
 	forceMove(stored_trap)//moves item into trap
 
 /obj/item/bountytrap/Destroy()
-	qdel(stored_trap)
+	if(stored_trap && !QDELING(stored_trap))
+		qdel(stored_trap)
+	stored_trap = null
 	QDEL_NULL(spark_system)
 	. = ..()
